LPS-55718 Moves EditRoleAction, no logic changes
